Pass dtype to zero valuet tensor in `_row_lengths_to_offsets`

+++ merlin/dataloader/jax.py
@@ -123,7 +123,7 @@ class Loader(LoaderBase):
         raise NotImplementedError("Sparse support isn't implemented yet for the Jax dataloader")
 
     def _row_lengths_to_offsets(self, row_lengths):
-        zero_value = jnp.array([0])
+        zero_value = jnp.array([0], dtype=row_lengths.dtype)
         if len(row_lengths.shape) == 2:
             zero_value = zero_value.reshape(-1, 1)
         return jnp.concatenate([zero_value, jnp.cumsum(row_lengths, axis=0)], axis=0)
